Fix profileImage for external or self-hosted profile pics
Does not yet work for viewing other user's profiles or contributions.
