+++ app/components/Header/UserDetails.js
@@ -1,39 +1,36 @@
-import React from 'react';
 import PropTypes from 'prop-types';
+import React, { useMemo } from 'react';
 import ReactTooltip from 'react-tooltip';
 
 import {
-  ProfilePic,
   EditGreen,
-  QuestionMarkGreen,
-  questionmark_orange
+  ProfilePic,
+  questionmark_orange,
+  QuestionMarkGreen
 } from '../../assets';
-import TextSpan from '../Common/Text/TextSpan';
-import TransparentButton from '../Common/Button/TransparentButton';
 import i18n from '../../locales/i18n.js';
-// import { getImageUrl } from '../../actions/apiRouting';
-import UserProfileImage from '../Common/UserProfileImage';
+import { profileImageUrl } from '../../utils/profileImage';
+import { useAuth0 } from '../auth0/react-auth0';
 import PrimaryButton from '../Common/Button/PrimaryButton';
+import TransparentButton from '../Common/Button/TransparentButton';
+import TextSpan from '../Common/Text/TextSpan';
 
-const UserDetails = ({
-  updateRoute,
-  userProfile,
-  logoutUser,
-  openProfilePickerModal
-}) => {
+const UserDetails = ({ updateRoute, userProfile, openProfilePickerModal }) => {
   const profile = userProfile || {};
+  const { user, logout } = useAuth0();
+
+  const logoutUser = () => logout({ redirect: window.location.origin });
+
+  const profileImage = useMemo(() => profileImageUrl(userProfile, user), [
+    userProfile,
+    user
+  ]);
+
   return (
     <div>
       <div className="popover__list-item">
         <div className="list-item__wrapper">
-          <UserProfileImage profileImage={profile.image} />
-          {/* <img
-            src={
-              userProfile.image
-                ? getImageUrl('profile', 'thumb', userProfile.image)
-                : ProfilePic
-            }
-          /> */}
+          <img src={profileImage} />
           <div>
             <TextSpan strong>
               {i18n.t('label.welcome_hi', {
@@ -69,7 +66,7 @@ const UserDetails = ({
             </div>
             <PrimaryButton
               className="pick-profile-primary-button"
-              onClick={() => openProfilePickerModal()}
+              onClick={openProfilePickerModal}
             >
               {i18n.t('label.edit')}
             </PrimaryButton>
@@ -78,7 +75,7 @@ const UserDetails = ({
           <div className="pick-profile-container">
             <PrimaryButton
               className="pick-profile-primary-button"
-              onClick={() => openProfilePickerModal()}
+              onClick={openProfilePickerModal}
             >
               {i18n.t('label.pick_profile')}
             </PrimaryButton>
@@ -104,9 +101,8 @@ const UserDetails = ({
 
 UserDetails.propTypes = {
   userProfile: PropTypes.object,
-  logoutUser: PropTypes.func.isRequired,
-  updateRoute: PropTypes.func,
-  openProfilePickerModal: PropTypes.func
+  updateRoute: PropTypes.func.isRequired,
+  openProfilePickerModal: PropTypes.func.isRequired
 };
 
 export default UserDetails;
